Adjusted if else statement and clean_complaint
Set wn_tag = None at first line, removing the else statement.
Set nlkt.tokenize in the functions (lemmatize_sentence & remove_non_alphabetica_char_and_x), as that will be used in the functions.
Lemmatizer is moved to the lemmatize_sentence function. It was unnecessary to set in the clean_complaint function